utils/requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
python-magic==0.4.24
mysql-connector-python==8.0.27